Move server code to seperate module
